对接最新feed分支，测试了用户注册、登录功能
