+++ controller/comment.go
@@ -1,11 +1,11 @@
 package controller
 
 import (
+	"douyin/dao"
+	"douyin/entity"
+	"douyin/pack"
+	"douyin/service"
 	"fmt"
-	"github.com/Usigned/douyin/dao"
-	"github.com/Usigned/douyin/entity"
-	"github.com/Usigned/douyin/pack"
-	"github.com/Usigned/douyin/service"
 	"github.com/gin-gonic/gin"
 	"net/http"
 	"strconv"
