--- pack/pack.go
@@ -1,17 +1,0 @@
-package pack
-
-// tools to pack models to entities
-// TODO
-
-import (
-	"github.com/Usigned/douyin/dao"
-	"github.com/Usigned/douyin/entity"
-)
-
-func PackUser(userModel *dao.User) *entity.User {
-	return nil
-}
-
-func MPackUser(userModels []*dao.User) []*entity.User {
-	return nil
-}
